+++ src/main/java/ListaExerciciosHerancaPolimorfismo/Ex4/VooInternacional.java
@@ -1,19 +1,19 @@
 package ListaExerciciosHerancaPolimorfismo.Ex4;
 
 public class VooInternacional extends Voo {
-    private int fatorPreco;
+    private float fatorPreco;
     private float taxaConversaoMoeda;
 
-    public VooInternacional(String origem, float distancia, String destino, String dataVoo, int fatorPreco){
+    public VooInternacional(String origem, float distancia, String destino, String dataVoo, float fatorPreco){
         super(origem, distancia, destino, dataVoo);
         setFatorPreco(fatorPreco);
     }
 
-    public int getFatorPreco() {
+    public float getFatorPreco() {
         return this.fatorPreco;
     }
 
-    public void setFatorPreco(int fatorPreco) {
+    public void setFatorPreco(float fatorPreco) {
         if(fatorPreco <= 0){
             throw new IllegalArgumentException("Erro: Argumento FatorPreco deve ser positivo");
         }
@@ -34,6 +34,4 @@ public class VooInternacional extends Voo {
     public float calcularPreco(){
         return getDistancia() * getFatorPreco();
     }
-
-    //TODO: casos de teste
 }
